[ros_distro_humble] Add explicit location of python3-numpy
We have the issue, that cmake will try to find python3 executable (which
is the python3-native package) but at times, we also need python3-numpy
includes from the python3-numpy (without -native) for the compilation.

We inject it here, otherwise we would have multiple recipes that would need
this extension.

Signed-off-by: Matthias Schoepfer <m.schoepfer@rethinkrobotics.com>